refactor(calendar): reorganize components into desktop/mobile/shared folders
