+++ frontend/app/components/calendar/desktop/CalendarNavbar.tsx
