+++ frontend/app/components/calendar/desktop/CalendarSidebar.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
-import TextButton from '../atoms/TextButton';
+import TextButton from '../../atoms/TextButton';
 
-import MiniCalendar from '../atoms/MiniCalendar';
-import MeetingsFilter from './MeetingsFilter';
-import { MeetingFilters } from '../../../util/meetingFilters';
-import NewMeetingSidebar from '../meeting-form/NewMeeting';
-import styles from '../../../styles/components/calendar/CalendarSidebar.module.scss';
+import MiniCalendar from '../../atoms/MiniCalendar';
+import MeetingsFilter from '../shared/MeetingsFilter';
+import { MeetingFilters } from '../../../../util/meetingFilters';
+import NewMeetingSidebar from '../../meeting-form/NewMeeting';
+import styles from '../../../../styles/components/calendar/desktop/CalendarSidebar.module.scss';
 import AddIcon from '@mui/icons-material/Add';
 interface CalendarSidebarProps {
   filters: MeetingFilters;
