+++ frontend/app/components/calendar/shared/MeetingsFilter.tsx
@@ -1,9 +1,9 @@
 "use client";
 
 import React from 'react';
-import FilterGroup, { FilterGroupItem } from '../shared/FilterGroup';
-import { ROOM_COLORS, ZOOM_ROOM_COLOR, CATEGORY_COLOR } from '../../../util/filterColors';
-import { MeetingFilters } from '../../../util/meetingFilters';
+import FilterGroup, { FilterGroupItem } from '../../shared/FilterGroup';
+import { ROOM_COLORS, ZOOM_ROOM_COLOR, CATEGORY_COLOR } from '../../../../util/filterColors';
+import { MeetingFilters } from '../../../../util/meetingFilters';
 
 interface MeetingsFilterProps {
     filters: MeetingFilters;
