+++ frontend/styles/components/calendar/desktop/CalendarNavbar.module.scss
