+++ frontend/styles/components/calendar/desktop/CalendarSidebar.module.scss
