+++ frontend/styles/components/calendar/desktop/CompactCalendarSidebar.module.scss
@@ -1,4 +1,4 @@
-@import '../../Variables.module.scss';
+@import '../../../Variables.module.scss';
 
 .rail {
   display: flex;
