--- frontend/styles/components/calendar/mobile/DayLandscapeSwitcher.module.scss
@@ -1,43 +1,0 @@
-@import '../../../Variables.module.scss';
-
-.container {
-  position: relative;
-  flex: 1 1 auto;
-  min-height: 0;
-  display: flex;
-  flex-direction: column;
-}
-
-// Floats over the calendar content rather than taking its own row -- landscape has ~350px
-// of height to work with total, not enough to spend a whole flex row on a view switcher.
-.viewToggle {
-  position: absolute;
-  top: 4px;
-  right: 4px;
-  z-index: $z-index-sticky;
-  display: flex;
-  gap: 2px;
-  padding: 2px;
-  background-color: rgb(255 255 255 / 92%);
-  border: 1px solid $grey-border-color;
-  border-radius: $radius-md;
-  box-shadow: 0 1px 3px rgb(0 0 0 / 15%);
-}
-
-.toggleButton {
-  appearance: none;
-  border: none;
-  background: none;
-  padding: 3px 8px;
-  border-radius: $radius-sm;
-  font-family: 'Source Sans Pro', sans-serif;
-  font-size: 10px;
-  font-weight: 600;
-  color: $text-secondary-color;
-  cursor: pointer;
-
-  &.active {
-    background-color: $black-color;
-    color: $white-color;
-  }
-}
