+++ frontend/styles/components/calendar/mobile/MobileFab.module.scss
@@ -1,4 +1,4 @@
-@import '../../Variables.module.scss';
+@import '../../../Variables.module.scss';
 
 .fabWrapper {
     position: fixed;
