+++ frontend/styles/components/calendar/shared/OverlapMeetingsModal.module.scss
@@ -1,5 +1,5 @@
 @import 'https://fonts.googleapis.com/css2?family=Source+Sans+Pro:wght@400;600&display=swap';
-@import '../../Variables.module.scss';
+@import '../../../Variables.module.scss';
 
 .modalOverlay {
   position: fixed;
